+++ lib/lambda_ethereum_consensus/state_transition/mutators.ex
@@ -8,20 +8,6 @@ defmodule LambdaEthereumConsensus.StateTransition.Mutators do
   alias SszTypes.Validator
 
   @doc """
-    Decrease the validator balance at index ``index`` by ``delta``, with underflow protection.
-  """
-  @spec decrease_balance(BeaconState.t(), SszTypes.validator_index(), SszTypes.gwei()) ::
-          BeaconState.t()
-  def decrease_balance(%BeaconState{balances: balances} = state, index, delta) do
-    current_balance = Enum.fetch!(balances, index)
-
-    %BeaconState{
-      state
-      | balances: List.replace_at(balances, index, max(current_balance - delta, 0))
-    }
-  end
-
-  @doc """
   Initiate the exit of the validator with index ``index``.
   """
   @spec initiate_validator_exit(BeaconState.t(), integer) ::
