+++ lib/ssz_types/beacon_chain/beacon_state.ex
@@ -88,4 +88,18 @@ defmodule SszTypes.BeaconState do
           # [New in Capella]
           historical_summaries: list(SszTypes.HistoricalSummary.t())
         }
+
+  @doc """
+    Decrease the validator balance at index ``index`` by ``delta``, with underflow protection.
+  """
+  @spec decrease_balance(BeaconState.t(), SszTypes.validator_index(), SszTypes.gwei()) ::
+          BeaconState.t()
+  def decrease_balance(%{balances: balances} = state, index, delta) do
+    current_balance = Enum.fetch!(balances, index)
+
+    %{
+      state
+      | balances: List.replace_at(balances, index, max(current_balance - delta, 0))
+    }
+  end
 end
